+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/driver/CronDriverProvider.java
@@ -1,5 +1,6 @@
 package com.dwarfeng.judge.impl.handler.driver;
 
+import com.dwarfeng.judge.impl.handler.DriverProvider;
 import com.dwarfeng.judge.stack.bean.entity.DriverInfo;
 import com.dwarfeng.judge.stack.exception.DriverException;
 import com.dwarfeng.judge.stack.handler.Driver;
@@ -9,62 +10,42 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
+import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.stereotype.Component;
 
 import java.util.HashSet;
+import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 /**
- * 固定间隔驱动提供器。
+ * Cron驱动提供器。
  *
  * @author DwArFeng
- * @since 1.3.0
+ * @since beta-1.0.0
  */
 @Component
-public class FixedRateDriverRegistry extends AbstractDriverRegistry {
+public class CronDriverProvider implements DriverProvider {
 
-    public static final String DRIVER_TYPE = "fixed_rate_driver";
+    public static final String SUPPORT_TYPE = "cron_driver";
 
     @Autowired
-    private FixedRateDriver fixedRateDriver;
-
-    public FixedRateDriverRegistry() {
-        super(DRIVER_TYPE);
-    }
+    private CronDriver cronDriver;
 
     @Override
-    public String provideLabel() {
-        return "固定频率驱动器";
-    }
-
-    @Override
-    public String provideDescription() {
-        return "根据指定的间隔定时驱动，如果某一次驱动晚于间隔，则后续驱动的时间会提前，以保持频率不变。";
-    }
-
-    @Override
-    public String provideExampleContent() {
-        return "60000";
+    public boolean supportType(String type) {
+        return Objects.equals(SUPPORT_TYPE, type);
     }
 
     @Override
     public Driver provide() {
-        return fixedRateDriver;
-    }
-
-    @Override
-    public String toString() {
-        return "FixedRateDriverRegistry{" +
-                "fixedRateDriver=" + fixedRateDriver +
-                ", driverType='" + driverType + '\'' +
-                '}';
+        return cronDriver;
     }
 
     @Component
-    public static class FixedRateDriver implements Driver {
+    public static class CronDriver implements Driver {
 
         @Autowired
         private ThreadPoolTaskScheduler scheduler;
@@ -73,21 +54,21 @@ public class FixedRateDriverRegistry extends AbstractDriverRegistry {
 
         private final Lock lock = new ReentrantLock();
         private final Set<ScheduledFuture<?>> scheduledFutures = new HashSet<>();
-        private final Set<FixedRateProcessor> fixedRateProcessors = new HashSet<>();
+        private final Set<CronProcessor> cronProcessors = new HashSet<>();
 
         @Override
         public void register(DriverInfo driverInfo) throws DriverException {
             lock.lock();
             try {
                 LongIdKey sectionKey = driverInfo.getSectionKey();
-                long rate = Long.parseLong(driverInfo.getContent());
-                FixedRateProcessor fixedRateProcessor = new FixedRateProcessor(
+                String cron = driverInfo.getContent();
+                CronProcessor cronProcessor = new CronProcessor(
                         evaluateService,
                         sectionKey
                 );
-                ScheduledFuture<?> scheduledFuture =
-                        scheduler.scheduleAtFixedRate(fixedRateProcessor, rate);
-                fixedRateProcessors.add(fixedRateProcessor);
+                CronTrigger cronTrigger = new CronTrigger(cron);
+                ScheduledFuture<?> scheduledFuture = scheduler.schedule(cronProcessor, cronTrigger);
+                cronProcessors.add(cronProcessor);
                 scheduledFutures.add(scheduledFuture);
             } catch (Exception e) {
                 throw new DriverException(e);
@@ -103,8 +84,8 @@ public class FixedRateDriverRegistry extends AbstractDriverRegistry {
                 for (ScheduledFuture<?> scheduledFuture : scheduledFutures) {
                     scheduledFuture.cancel(true);
                 }
-                for (FixedRateProcessor fixedRateProcessor : fixedRateProcessors) {
-                    fixedRateProcessor.shutdown();
+                for (CronProcessor cronProcessor : cronProcessors) {
+                    cronProcessor.shutdown();
                 }
             } finally {
                 lock.unlock();
@@ -112,9 +93,9 @@ public class FixedRateDriverRegistry extends AbstractDriverRegistry {
         }
     }
 
-    private static class FixedRateProcessor implements Runnable {
+    private static class CronProcessor implements Runnable {
 
-        private static final Logger LOGGER = LoggerFactory.getLogger(FixedRateProcessor.class);
+        private static final Logger LOGGER = LoggerFactory.getLogger(CronProcessor.class);
 
         private final EvaluateService evaluateService;
         private final LongIdKey sectionKey;
@@ -122,7 +103,7 @@ public class FixedRateDriverRegistry extends AbstractDriverRegistry {
         private final Lock lock = new ReentrantLock();
         private boolean runningFlag = true;
 
-        private FixedRateProcessor(EvaluateService evaluateService, LongIdKey sectionKey) {
+        private CronProcessor(EvaluateService evaluateService, LongIdKey sectionKey) {
             this.evaluateService = evaluateService;
             this.sectionKey = sectionKey;
         }
@@ -135,7 +116,7 @@ public class FixedRateDriverRegistry extends AbstractDriverRegistry {
                     return;
                 }
 
-                LOGGER.debug("计划时间已到达, fixed rate 驱动器驱动 " + sectionKey + " 部件执行评估动作...");
+                LOGGER.debug("计划时间已到达, cron驱动器驱动 " + sectionKey + " 部件执行评估动作...");
                 evaluateService.evaluate(sectionKey);
             } catch (Exception e) {
                 LOGGER.warn("记录 " + sectionKey + " 时出现异常, 放弃本次记录", e);